+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 CHANGELOG
 =========
 
+v0.1.2
+------
+
+- Use the default config file as a fallback if options are missing in user config file
+- Do not show previously shown notifications
+- Add dbus-python as a dependency for linux
+- Properly handle Control-C
+
 v0.1.1
 ------